feat(domain): services return Result<T> and use LogService; migrate IDs to int

+++ lib/features/projects/presentation/screens/project_detail_screen.dart
@@ -19,7 +19,7 @@ import '../providers/project_provider.dart';
 import '../widgets/project_detail_header.dart';
 
 class ProjectDetailScreen extends ConsumerStatefulWidget {
-  final BigInt projectId;
+  final int projectId;
 
   const ProjectDetailScreen({super.key, required this.projectId});
