+++ lib/features/tasks/data/models/task_model.dart
@@ -9,11 +9,11 @@ import '../../domain/entities/task_priority.dart';
 @TableIndex(name: 'task_completed_idx', columns: {#isCompleted})
 @TableIndex(name: 'task_updated_at_idx', columns: {#updatedAt})
 class TaskTable extends Table {
-  Int64Column get id => int64().autoIncrement()();
+  IntColumn get id => integer().autoIncrement()();
 
-  Int64Column get projectId => int64()();
+  IntColumn get projectId => integer()();
 
-  Int64Column get parentTaskId => int64().nullable().references(TaskTable, #id)();
+  IntColumn get parentTaskId => integer().nullable().references(TaskTable, #id)();
 
   TextColumn get title => text()();
